Fixes #13

+++ Sketches/ESP-IDF/ControlAndInterface/src/DataTypes.h
@@ -10,5 +10,5 @@ typedef int32_t Velocity;
 typedef int16_t Turns;
 typedef int8_t Torque;
 typedef uint16_t Current;
-typedef int16_t Period; // us
-typedef int16_t Frequency; // Hz
+typedef int32_t Period; // us
+typedef int32_t Frequency; // Hz
